Select Flows That Cross Nodes button now moves flows off of nodes.
This has a few issues. For one, it moves the control point
perpendicularly away from the flow’s baseline, no matter where the
offending nodes are. This is not always optimal. Also, it will continue
to move the control point in this direction until it no longer
intersects a node. This is not always possible, and can easily result
in an infinite loop.